% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLmatrix.R
\name{CLmatrix}
\alias{CLmatrix}
\title{The (positive) Chow-Liu matrix}
\usage{
CLmatrix(S, positive = TRUE)
}
\arguments{
\item{S}{a sample covariance matrix}

\item{positive}{TRUE}
}
\description{
Computes the Chow-Liu estimate or the positive Chow-Liu estimate in which 
case only nonnnegative entries of S are taken into account. The Chow-Liu estimate 
corresponds to the maximum likelihood over all Gaussian tree models.
}
\examples{
S <- matrix(c(11.376,0.745,8.446,2.066,7.686, -0.759,0.745, 33.069,0.672, 
36.047,2.016, 32.059,8.446,0.672,8.937,0.315,6.864, -0.598,2.066, 36.047,
0.315, 51.939,2.187, 41.587,7.686,2.016,6.864,2.187,7.645,0.384, -0.759, 
32.059, -0.598, 41.587,0.384, 41.560),6,6)
C <- CLmatrix(S)
}
\keyword{Chow-Liu}
\keyword{algorithm,}
\keyword{cost}
\keyword{minimum}
\keyword{spanning}
\keyword{tree}
