% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLsigns.R
\name{CLsigns}
\alias{CLsigns}
\title{The Chow-Liu sign adjustment}
\usage{
CLsigns(S)
}
\arguments{
\item{S}{a sample covariance matrix}
}
\description{
omputes the Chow-Liu estimate or the positive Chow-Liu estimate in which 
case only nonnnegative entries of S are taken into account. The Chow-Liu estimate 
corresponds to the maximum likelihood over all Gaussian tree models.
}
\examples{



}
\keyword{Chow-Liu}
\keyword{algorithm,}
\keyword{cost}
\keyword{minimum}
\keyword{spanning}
\keyword{tree}

