% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord2K.R
\name{coord2K}
\alias{coord2K}
\title{A coordinate descent algorithm to find the maximum likelihood estimator by local updates of the concntration matrix K.}
\usage{
coord2K(S, n = 1, K0 = solve(diag(diag(S))), tol = 1e-08)
}
\arguments{
\item{S}{the sample covariance matrix}

\item{n}{the sample size (default 1)}

\item{K0}{the starting point (default K0=solve(diag(diag(S))))}

\item{tol}{the convergence tolerance (default tol=1e-8)}
}
\value{
the optimal value of the concentration matrix

the number of iterations the algorithm needed to converge

the corresponding value of the log-likelihood
}
\description{
This function implements a simple coordinate descent algorithm to find the maximum likelihood estimator over
Gaussian MTP2 distributions. For details see Lauritzen, Uhler, Zwiernik (2017).
}
\examples{


}
\keyword{concentration}
\keyword{coordinate}
\keyword{descent,}
\keyword{matrix.}

