% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord2Sig.R
\name{coord2Sig}
\alias{coord2Sig}
\title{Coordinate descent algorithm for the MLE.}
\usage{
coord2Sig(S, n = 1, Sig0 = S, tol = 1e-10, graph = FALSE)
}
\arguments{
\item{S}{the sample covariance matrix}

\item{n}{the sample size (default 1)}

\item{Sig0}{starting point of the algorithm}

\item{tol}{convergence criterion sum(abs(Sig-Sig0))<tol}
}
\description{
This function implements a simple coordinate descent algorithm to find the maximum likelihood estimator over
Gaussian MTP2 distributions. The algorithm performs local updates of the covariance matrix $Sigma$.
For details see Lauritzen, Uhler, Zwiernik (2017); look for Algorithm 4.
}
\examples{
library(gRbase)
data("carcass")
S <- cov(carcass)[1:6,1:6]
coord2Sig(S)

}
\keyword{coordinate}
\keyword{covariance}
\keyword{descent,}
\keyword{matrix.}

