% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBICgolazo.R
\name{EBICgolazo}
\alias{EBICgolazo}
\title{Finds  the optimal penalty parameter for Graphical Oriented Lasso using the EBIC criterion.}
\usage{
EBICgolazo(
  S,
  n = NULL,
  L = NULL,
  U = NULL,
  tol = 1e-06,
  edge.tol = 1e-04,
  gamma = 0.5,
  rhomin = 0.01,
  rhomax = 1,
  nrhos = 50,
  output = TRUE
)
}
\arguments{
\item{S}{the sample covariance matrix}

\item{n}{the sample size}

\item{L}{matrix of lower penalties (can be -Inf)}

\item{U}{matrix of upper penalties (can be Inf)}

\item{tol}{the convergence tolerance (default tol=1e-8)}

\item{edge.tol}{which entries of K are treated as zero}

\item{gamma}{the EBIC parameter between 0 and 1}

\item{rhomin}{minimal rho to check}

\item{rhomax}{maximal rho to check}

\item{nrhos}{positive penalty (can be Inf)}

\item{output}{if TRUE (default) the output will be printed.}
}
\value{
the optimal value of the concentration matrix

the number of iterations the algorithm needed to converge

the corresponding value of the log-likelihood
}
\description{
This function computes the EBIC criterion for a seried of rho parameters. The penalty matrices are $rho L$, $rho U$, 
where L and U are fixed in  advance.
}
\examples{
print(TRUE)

}
\keyword{concentration}
\keyword{coordinate}
\keyword{descent,}
\keyword{matrix.}
