% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPlasso.R
\name{EPlasso}
\alias{EPlasso}
\title{Performs l1-regularized estimation of an inverse covariance matrix assuming nonnegativity of correlations}
\usage{
EPlasso(S, lambda, tol = 1e-07, pos.constr = TRUE)
}
\arguments{
\item{S}{the sample covariance matrix}

\item{lambda}{positive penalty}

\item{tol}{the convergence tolerance (default tol=1e-8)}

\item{pos.constr}{if TRUE (default) assumes nnonnegative correlations, if FALSE performs the standard dual graphical lasso.}
}
\value{
the optimal value of the concentration matrix

the number of iterations the algorithm needed to converge

the corresponding value of the log-likelihood
}
\description{
This function implements a simple coordinate descent algorithm to find the maximum likelihood estimator over
Gaussian MTP2 distributions. For details see Lauritzen, Uhler, Zwiernik (2017).
}
\examples{


}
\keyword{concentration}
\keyword{coordinate}
\keyword{descent,}
\keyword{matrix.}
