% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_M.R
\name{is.M}
\alias{is.M}
\title{Checks if a positive definite matrix is an M-matrix}
\usage{
is.M(K, tol = 1e-10)
}
\arguments{
\item{K}{an inverse covariance matrix}

\item{tol}{numbers greater than tol are treated as positive}
}
\description{
Checks if all entries of K are smaller than a small positive number tol
}
\examples{
print(TRUE)

}
\keyword{M-matrix}
