% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LADualMLE.R
\name{LADualMLE}
\alias{LADualMLE}
\title{A coordinate descent algorithm to find the dual maximum likelihood estimator for locally associated Gaussian distributions}
\usage{
LADualMLE(W, tol = 1e-08)
}
\arguments{
\item{W}{inverse of the sample covariance matrix ???}

\item{tol}{the convergence tolerance (default tol=1e-8)}
}
\value{
the optimal value of the concentration matrix

the number of iterations the algorithm needed to converge

the corresponding value of the log-likelihood
}
\description{
This function implements a simple coordinate descent algorithm to find the maximum likelihood estimator over
Gaussian MTP2 distributions. For details see Lauritzen, Uhler, Zwiernik (2017).
}
\examples{
print(TRUE)

}
\keyword{concentration}
\keyword{coordinate}
\keyword{descent,}
\keyword{matrix.}
