% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP.simplicial.complex.R
\name{LP.simplicial.complex}
\alias{LP.simplicial.complex}
\title{Locally positive simplicial complex of a covariance matrix.
This function gives the set of all subblocks C for which Sigma[C,C] is an inverse M-matrix}
\usage{
LP.simplicial.complex(S, tol = 1e-07)
}
\arguments{
\item{S}{the sample covariance matrix}

\item{tol}{the convergence tolerance (default tol=1e-8)}
}
\value{
the simplicial complex
}
\description{
Locally positive simplicial complex of a covariance matrix.
This function gives the set of all subblocks C for which Sigma[C,C] is an inverse M-matrix
}
\examples{


}
\keyword{M-subblocks,}
\keyword{complex,}
\keyword{inverse}
\keyword{local}
\keyword{positivity}
\keyword{simplicial}
