% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPFit.R
\name{MPFit}
\alias{MPFit}
\title{A coordinate descent algorithm to find the maximum likelihood estimator for Markov positive distribution by local updates .}
\usage{
MPFit(cliques, S, n = 1, tol = 1e-08)
}
\arguments{
\item{cliques}{the list ov vectors with indices of cliques}

\item{S}{the sample covariance matrix}

\item{n}{the sample size (default 1), relevant for testing}

\item{tol}{the convergence tolerance (default tol=1e-8)}
}
\value{
the optimal value of the concentration matrix

the number of iterations the algorithm needed to converge

the corresponding value of the log-likelihood
}
\description{
This function implements a simple coordinate descent algorithm to find the maximum likelihood estimator over
Gaussian MTP2 distributions. For details see Lauritzen, Uhler, Zwiernik (2017).
}
\examples{


}
\keyword{concentration}
\keyword{coordinate}
\keyword{descent,}
\keyword{matrix.}
