% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpLR.R
\name{mtpLR}
\alias{mtpLR}
\title{The likelihood ratio test for MTP2}
\usage{
mtpLR(K, LR, tol = 1e-10)
}
\arguments{
\item{K}{an M-matrix that maximizes the likelihood}

\item{LR}{the value of the likelihood ratio statistics with respect to the unconstrained model}

\item{tol}{numbers greater than -tol in K are treated as zero (default 1e-8)}

\item{n}{the sample size}
}
\description{
Internal funciton that computes the likelihood ratio statistics to test the MTP constraints. It also returns the degrees 
of freedom and the corresponding p-value.
}
\examples{

}
\keyword{xxx}

