% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWSF.R
\name{MWSF}
\alias{MWSF}
\title{The (positive) maximum weight spanning forest}
\usage{
MWSF(S, positive = TRUE)
}
\arguments{
\item{S}{a sample covariance matrix}

\item{positive}{TRUE}
}
\description{
Computes the Chow-Liu estimate or the positive Chow-Liu estimate in which 
case only nonnnegative entries of S are taken into account. The Chow-Liu estimate 
corresponds to the maximum likelihood over all Gaussian tree models.
}
\examples{
library(gRbase)
data(carcass)
S <- cov(carcass)[1:6,1:6]
C <- CLmatrix(S)
graphK(solve(C))
}
\keyword{Chow-Liu}
\keyword{algorithm,}
\keyword{cost}
\keyword{minimum}
\keyword{spanning}
\keyword{tree}

