% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.1.R
\name{skeleton.1}
\alias{skeleton.1}
\title{Compute the 1-skeleton of a simplicial complex}
\usage{
skeleton.1(simplices)
}
\arguments{
\item{simplices}{the list ov maximal simplices}
}
\value{
the optimal value of the concentration matrix

the number of iterations the algorithm needed to converge

the corresponding value of the log-likelihood
}
\description{
This function implements a simple coordinate descent algorithm to find the maximum likelihood estimator over
Gaussian MTP2 distributions. For details see Lauritzen, Uhler, Zwiernik (2017).
}
\examples{
print(TRUE)

}
\keyword{concentration}
\keyword{coordinate}
\keyword{descent,}
\keyword{matrix.}
