% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wmatrix.R
\name{Wmatrix}
\alias{Wmatrix}
\title{The W matrix}
\usage{
Wmatrix(S)
}
\arguments{
\item{S}{a sample covariance matrix}
}
\description{
Computes the W matrix as introduced in Section 4.2 of Lauritzen, Uhler, Zwiernik (2017).
}
\examples{


}
\keyword{W}
\keyword{matrix}
\keyword{path,}
\keyword{shortest}
