% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zmatrix.R
\name{Zmatrix}
\alias{Zmatrix}
\title{The graphs of an M-matrix}
\usage{
Zmatrix(S)
}
\arguments{
\item{K}{an M-matrix}

\item{tol}{numbers greater than -tol are treated as zero}
}
\description{
This function oututs the Z matrix, that is, the unique dominating ultrametric.
}
\examples{

}
\keyword{xxx}

