% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphK.R
\name{graphK}
\alias{graphK}
\title{The graph of an M-matrix}
\usage{
graphK(S, K, names = FALSE, tol = 1e-10, root = c())
}
\arguments{
\item{S}{the sample covariance matrix}

\item{K}{an M-matrix}

\item{tol}{numbers greater than -tol are treated as zero}
}
\description{
Plots the graph representing the support of an M-matrix. 
Small negative numbers are treated as zero.
}
\examples{

}
\keyword{xxx}

