% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R, R/is_M.R
\name{is.M}
\alias{is.M}
\title{The graph of an M-matrix}
\usage{
is.M(K, tol = 1e-10)

is.M(K, tol = 1e-10)
}
\arguments{
\item{K}{an M-matrix}

\item{tol}{numbers greater than -tol are treated as zero}

\item{S}{the sample covariance matrix}

\item{S}{the sample covariance matrix}

\item{K}{an M-matrix}

\item{tol}{numbers greater than -tol are treated as zero}
}
\description{
Plots the graph representing the support of an M-matrix. 
Small negative numbers are treated as zero.

Plots the graph representing the support of an M-matrix. 
Small negative numbers are treated as zero.
}
\examples{




}
\keyword{xxx}

