% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zmatrix.R
\name{Zmatrix}
\alias{Zmatrix}
\title{Computes the Z-matrix}
\usage{
Zmatrix(S)
}
\arguments{
\item{S}{a covariance matrix}
}
\description{
This function oututs the Z matrix, that is, the unique dominating ultrametric.
}
\examples{

}
\keyword{xxx}
